requirements.txt
pillow
ollama
typer[all]
requests
